+++ packages/cli/src/__tests__/test-helpers.ts
@@ -175,6 +175,61 @@ export function mockClackPrompts(overrides?: Partial<ClackPromptsMock>): ClackPr
   return mocks;
 }
 
+// ── Bun.spawn Mock ─────────────────────────────────────────────────────────────
+
+/**
+ * Mocks Bun.spawn to return a fake process with the given exit code, stdout, and stderr.
+ * Identical helper was previously duplicated across aws-cov, gcp-cov, do-cov, hetzner-cov,
+ * and sprite-cov test files. Centralised here to avoid repetition.
+ */
+export function mockBunSpawn(exitCode = 0, stdout = "", stderr = "") {
+  const mockProc = {
+    pid: 1234,
+    exitCode: Promise.resolve(exitCode),
+    exited: Promise.resolve(exitCode),
+    stdout: new ReadableStream({
+      start(c) {
+        c.enqueue(new TextEncoder().encode(stdout));
+        c.close();
+      },
+    }),
+    stderr: new ReadableStream({
+      start(c) {
+        c.enqueue(new TextEncoder().encode(stderr));
+        c.close();
+      },
+    }),
+    kill: mock(() => {}),
+    ref: () => {},
+    unref: () => {},
+    stdin: new WritableStream(),
+    resourceUsage: () =>
+      ({
+        cpuTime: {
+          system: 0,
+          user: 0,
+          total: 0,
+        },
+        maxRSS: 0,
+        sharedMemorySize: 0,
+        unsharedDataSize: 0,
+        unsharedStackSize: 0,
+        minorPageFaults: 0,
+        majorPageFaults: 0,
+        swapCount: 0,
+        inBlock: 0,
+        outBlock: 0,
+        ipcMessagesSent: 0,
+        ipcMessagesReceived: 0,
+        signalsReceived: 0,
+        voluntaryContextSwitches: 0,
+        involuntaryContextSwitches: 0,
+      }) satisfies ReturnType<ReturnType<typeof Bun.spawn>["resourceUsage"]>,
+  };
+  // biome-ignore lint: test mock
+  return spyOn(Bun, "spawn").mockReturnValue(mockProc as ReturnType<typeof Bun.spawn>);
+}
+
 // ── Fetch Mocks ────────────────────────────────────────────────────────────────
 
 export function mockSuccessfulFetch(data: unknown) {
